extract agent file from smr
